enhanced AtlasManager
* saveResults() now throws an error if the atlas hierarchy contains regions classified as "Exclude"
* isSplit() added
* fixExclusions() added
* checks added on the atlas hierarchy before many operations are performed

+++ src/main/java/qupath/ext/braian/AtlasManager.java
@@ -9,6 +9,7 @@ import qupath.lib.common.GeneralTools;
 import qupath.lib.gui.measure.ObservableMeasurementTableData;
 import qupath.lib.objects.PathAnnotationObject;
 import qupath.lib.objects.PathObject;
+import qupath.lib.objects.PathObjectTools;
 import qupath.lib.objects.classes.PathClass;
 import qupath.lib.objects.hierarchy.PathObjectHierarchy;
 import qupath.lib.projects.ProjectImageEntry;
@@ -25,11 +26,24 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static qupath.ext.braian.BraiAnExtension.getLogger;
-import static qupath.ext.braian.BraiAnExtension.logger;
 
 class ImportedAtlasNotFound extends RuntimeException {
     public ImportedAtlasNotFound() {
         super("No previously imported atlas found. Use ABBA to align the slice to an atlas and import the brain annotations in the QuPath project ");
+    }
+}
+
+class DisruptedAtlasHierarchy extends RuntimeException {
+    public DisruptedAtlasHierarchy(PathObject atlas) {
+        super("The atlas hierarchy '"+atlas+"' was disrupted. "+
+                "You must import the atlas annotations again and delete any previous region annotation.");
+    }
+}
+
+class ExclusionMistakeException extends RuntimeException {
+    public ExclusionMistakeException() {
+        super("Some regions in the atlas ontology were wrongly classified as '"+AtlasManager.EXCLUDE_CLASSIFICATION+"'.\n"+
+                "You can try to fix this by calling fixExclusions() on the AtlasManager instance.");
     }
 }
 
@@ -39,6 +53,8 @@ class ImportedAtlasNotFound extends RuntimeException {
 public class AtlasManager {
     public final static String um = GeneralTools.micrometerSymbol();
     public final static PathClass EXCLUDE_CLASSIFICATION = PathClass.fromString("Exclude");
+    public final static PathClass ABBA_LEFT = PathClass.fromString("Left");
+    public final static PathClass ABBA_RIGHT = PathClass.fromString("Right");
 
     /**
      * Checks whether an ABBA atlas was previously imported.
@@ -82,7 +98,7 @@ public class AtlasManager {
 
     private static List<PathObject> getExclusionAnnotations(PathObjectHierarchy hierarchy) {
         return hierarchy.getAnnotationObjects().stream()
-                .filter( ann -> AtlasManager.EXCLUDE_CLASSIFICATION.equals(ann.getPathClass()) )
+                .filter( ann -> ann.getPathClass() == AtlasManager.EXCLUDE_CLASSIFICATION )
                 .toList();
     }
 
@@ -91,6 +107,7 @@ public class AtlasManager {
 
     /**
      * @param hierarchy where to search for the atlas
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
      */
     public AtlasManager(PathObjectHierarchy hierarchy) {
         this.hierarchy = hierarchy;
@@ -98,6 +115,8 @@ public class AtlasManager {
         if (atlases.isEmpty())
             throw new ImportedAtlasNotFound();
         this.atlasObject = atlases.get(0);
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
         if (atlases.size()>1)
             getLogger().warn("Several imported atlases have been found. Selecting: {}", this.atlasObject);
     }
@@ -106,9 +125,12 @@ public class AtlasManager {
      * Flattens the atlas's ontology into a list of annotations.
      * It may return some non-region annotations too, if the atlas hierarchy was modified with added/removed elements.
      * @return the list of all brain regions in the atlas
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
      * @see #flatten(List)
      */
     public List<PathObject> flatten() {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
         return AtlasManager.flattenObject(this.atlasObject);
     }
 
@@ -119,9 +141,12 @@ public class AtlasManager {
      * <br>
      * It may still return some non-region annotations, if the atlas hierarchy was further modified with added/removed elements.
      * @return the list of all brain regions in the atlas
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
      * @see #flatten()
      */
     public List<PathObject> flatten(List<AbstractDetections> detections) {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
         List<PathAnnotationObject> containers = detections.stream().flatMap(d -> d.getContainers().stream()).toList();
         List<PathObject> brainRegions = AtlasManager.flattenObjectStream(this.atlasObject)
                 .filter(ann -> !containers.contains(ann))
@@ -135,9 +160,13 @@ public class AtlasManager {
      * The table is saved as a CSV (comma-separated values) file if 'file' ends with ".csv"
      * @param detections the list of detection of which to gather the data for each region
      * @param file the file where it should write to. Note that if the file exists, it will be overwritten
+     * @throws ExclusionMistakeException if the atlas hierarchy contains regions classified as {@link #EXCLUDE_CLASSIFICATION}.
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
      */
     // Olivier Burri <https://github.com/lacan> wrote mostly of this function and published under Apache-2.0 license for qupath-extension-biop
     public boolean saveResults(List<AbstractDetections> detections, File file) {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
         if (file.exists())
             if(!file.delete()) {
                 getLogger().error("Could not delete previous results file {}, the file could be locked.", file.getName());
@@ -150,6 +179,8 @@ public class AtlasManager {
 
         ObservableMeasurementTableData ob = new ObservableMeasurementTableData();
         List<PathObject> brainRegions = this.flatten(detections);
+        if (brainRegions.stream().anyMatch(p -> p.getPathClass() == EXCLUDE_CLASSIFICATION))
+            throw new ExclusionMistakeException();
         // This line creates all the measurements
         ob.setImageData(QP.getCurrentImageData(), brainRegions);
 
@@ -196,15 +227,18 @@ public class AtlasManager {
      *   <li>either be contained into a larger annotation, classified as "Exclude"
      * 	 <li>or be <b>duplicated</b> (SHIFT+D) outside the Atlas's hierarchy, and then classified as "Exclude"
      * </ul>
-     * @return a set of brain regions' annotations that should be excluded
+     * @return a set of brain regions' annotations that should be excluded.
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
      */
     public Set<PathObject> getExcludedBrainRegions() {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
         List<PathObject> excludeAnnotations = AtlasManager.getExclusionAnnotations(this.hierarchy);
         getLogger().info("Exclusion annotations: [{}]", BraiAn.join(excludeAnnotations, ", "));
         // We export all the annotations that are not "Exclude" or named "Root".
         // This serves as a downstream ~check~ that the slice has nothing else other than the atlas annotations and the exclusions
         var otherAnnotations = this.hierarchy.getAnnotationObjects().stream()
-                .filter( ann -> !(AtlasManager.EXCLUDE_CLASSIFICATION.equals(ann.getPathClass()) || ann.equals(this.atlasObject)))
+                .filter( ann -> ann.getPathClass() != AtlasManager.EXCLUDE_CLASSIFICATION && ann != this.atlasObject)
                 .toList();
         // Loop over exclusions that contain the annotations to be removed/excluded
         Set<PathObject> regionsToExcludeFlattened = excludeAnnotations.stream()
@@ -225,10 +259,14 @@ public class AtlasManager {
     /**
      * Saves a file containing, on each line, the name and hemisphere of the regions to be excluded.
      * @param file the file where it should write to. Note that if the file exists, it will be overwritten
+     * @return true if the file was correctly saved.
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
      * @see #getExcludedBrainRegions()
      * @see #saveResults(List, File)
      */
     public boolean saveExcludedRegions(File file) {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
         if (file.exists())
             if(!file.delete()) {
                 getLogger().error("Could not delete previous exclusion file {}, the file could be locked.", file.getName());
@@ -256,5 +294,97 @@ public class AtlasManager {
         }
         getLogger().info("Exclusions '{}' Saved under '{}', contains {} rows", file.getName(), file.getParentFile().getAbsolutePath(), regionsToExclude.size());
         return true;
+    }
+
+    /**
+     * Return whether the current atlas is split between left and right hemispheres.
+     * Atlas hierarchies that were modified by deleting one of the two hemispheres, are still recognised as split.
+     * @return true if the current atlas is split between left and right hemispheres.
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
+     * whether the atlas was split between left and right.
+     */
+    public boolean isSplit() {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
+        List<PathObject> roots = this.atlasObject.getChildObjects().stream()
+                .filter(ann -> "root".equals(ann.getName()))
+                .toList();
+        return roots.stream()
+                // anyMatch because if there are multiple "root" annotations, an
+                .anyMatch(root -> {
+                            Set<PathClass> hemispheres = flattenObject(root).stream()
+                                    .map(PathObject::getPathClass)
+                                    .filter(c -> c.isDerivedFrom(ABBA_LEFT) || c.isDerivedFrom(ABBA_RIGHT))
+                                    .map(PathClass::getParentClass)
+                                    .collect(Collectors.toSet());
+                            if (hemispheres.size() == 2)
+                                throw new DisruptedAtlasHierarchy(this.atlasObject);
+                            return hemispheres.size() == 1;
+                        }
+                );
+    }
+
+    /**
+     * Attempts to fix possible mistakes with the exclusions, such as:
+     *  <ul>
+     *   <li>Exclusion of the "Root" hierarchy annotation rather than a proper region</li>
+     *   <li>If a region was excluded within the atlas hierarchy</li>
+     * </ul>
+     * @throws DisruptedAtlasHierarchy if the current atlas hierarchy was disrupted, and it is not possible to deduce
+     */
+    public void fixExclusions() {
+        if (this.atlasObject.getChildObjects().isEmpty())
+            throw new DisruptedAtlasHierarchy(this.atlasObject);
+        if (this.atlasObject.getPathClass() == AtlasManager.EXCLUDE_CLASSIFICATION) {
+            for (PathObject root: this.atlasObject.getChildObjects()) {
+                PathObject duplicate = PathObjectTools.transformObject(root, null, true, true);
+                duplicate.setPathClass(AtlasManager.EXCLUDE_CLASSIFICATION);
+                this.hierarchy.addObject(duplicate);
+            }
+            this.atlasObject.setPathClass(null); // TODO: set the proper path class, like ABBA 0.3.2 does
+        }
+        this.flatten().stream()
+                .filter(
+                        region -> region.getPathClass() != null && region.getPathClass() == AtlasManager.EXCLUDE_CLASSIFICATION
+                ).forEach(
+                        mistakenlyExcludedRegion -> this.fixMistakenlyExcludedRegion(mistakenlyExcludedRegion, this.isSplit())
+                );
+    }
+
+    private void fixMistakenlyExcludedRegion(PathObject mistakenlyExcludedRegion, boolean isSplit) {
+        String regionName;
+        if ((regionName = mistakenlyExcludedRegion.getName()) == null)
+            throw new RuntimeException("Can't deduce the name for brain region '"+mistakenlyExcludedRegion+"'.");
+        PathClass classification;
+        if (isSplit)
+            classification = PathClass.getInstance(this.getHemisphereOfMisclassifiedRegion(mistakenlyExcludedRegion),  regionName, null);
+        else
+            classification = PathClass.fromString(regionName);
+        List<PathObject> wrongDuplicates = this.hierarchy.getRootObject().getChildObjects().stream().filter(
+            ann -> regionName.equals(ann.getName()) && ann.getPathClass() == classification
+        ).toList();
+        if (wrongDuplicates.isEmpty()) {
+            PathObject duplicate = PathObjectTools.transformObject(mistakenlyExcludedRegion, null, true, true);
+            this.hierarchy.addObject(duplicate);
+            wrongDuplicates = List.of(duplicate);
+        }
+        for (PathObject wrongDuplicate: wrongDuplicates) {
+            wrongDuplicate.setPathClass(AtlasManager.EXCLUDE_CLASSIFICATION);
+        }
+        mistakenlyExcludedRegion.setPathClass(classification);
+    }
+
+    private PathClass getHemisphereOfMisclassifiedRegion(PathObject region) {
+        PathObject hemisphereObject = region;
+        while (hemisphereObject.getParent() != this.atlasObject)
+            hemisphereObject = hemisphereObject.getParent();
+        Set<PathClass> hemisphere = flattenObject(hemisphereObject).stream()
+                .map(PathObject::getPathClass)
+                .filter(c -> c.isDerivedClass() && (c.isDerivedFrom(ABBA_LEFT) || c.isDerivedFrom(ABBA_RIGHT)))
+                .map(PathClass::getParentClass)
+                .collect(Collectors.toSet());
+        if (hemisphere.size() != 1)
+            throw new RuntimeException("Can't deduce the hemisphere for '"+region+"'.");
+        return hemisphere.iterator().next();
     }
 }
